Looking better!

+++ components/chartView/blockContent.js
@@ -1,23 +1,30 @@
 import React, { useState } from "react";
 import { shiftHue } from "../utils/colors";
 import { deleteDoc, useDocument } from "swr-firestore-v9";
+import { DocumentContext } from "./documentContext";
 
 export default function BlockContent({
   item,
   hue = shiftHue(0, 4000, item.amount, 0),
   isRemainder = false,
-  hasRemainder = false,
+  remainder = null,
   isIncome = false,
   isOverBudget,
   hasChildren = false,
   handleAddChild,
+  isReadOnly,
 }) {
   const documentPath = `documents/UEXue3UyAZM8SvFyBhZP/items/${item.id}`;
+  const docContext = React.useContext(DocumentContext);
   const handleDelete = () => {
+    if (isIncome) {
+      const incomeItems = docContext.items.filter((x) => x.isIncome === true);
+      if (incomeItems.length <= 1) return;
+    }
     deleteDoc(documentPath);
   };
-  const { update } = useDocument(documentPath);
 
+  const { update } = useDocument(documentPath);
   const handleUpdateValue = (type, value) => {
     console.log("updating value to ", value);
     if (!item.id) return; // Exclude 'all income' block
@@ -34,36 +41,45 @@ export default function BlockContent({
 
   return (
     <div
-      className={`bg-gray-100 w-56 p-2 group flex flex-col justify-between text-right items-end ${
+      className={`bg-gray-100 w-56 p-2 group flex flex-col relative justify-between text-right items-end ${
         isRemainder && "text-gray-400"
       } ${isOverBudget && "text-red-600"}`}
       style={{
         textAlign: "right",
-        background: `hsla(${hue}, 80%, 70%, 1)`,
+        background: remainder >= 0 ? `hsla(${hue}, 80%, 70%, 1)` : `red`,
         borderTopRightRadius: !isIncome && !hasChildren ? 32 : 0,
         borderBottomRightRadius:
-          !isIncome && (!hasChildren || hasRemainder) ? 32 : 0,
+          !isIncome && (!hasChildren || remainder > 0) ? 32 : 0,
         borderTopLeftRadius: isIncome && 32,
         borderBottomLeftRadius: isIncome && 32,
       }}
     >
       <div className="max-w-48 overflow-hidden">
-        <Input
-          type="text"
-          initialValue={item.title}
-          handleBlur={handleUpdateValue}
-        />
-        <p className="font-semibold text-xl opacity-70 bg-blend-overlay text-black">
-          £
+        {isReadOnly ? (
+          <p className="font-semibold text-xl opacity-70 bg-blend-overlay text-black">
+            {item.title}
+          </p>
+        ) : (
           <Input
-            type="number"
-            initialValue={Math.round(item.amount)}
+            type="text"
+            initialValue={item.title}
             handleBlur={handleUpdateValue}
           />
+        )}
+        <p className="font-semibold text-xl opacity-70 bg-blend-overlay text-black">
+          {isReadOnly ? (
+            Math.round(item.amount)
+          ) : (
+            <Input
+              type="number"
+              initialValue={Math.round(item.amount)}
+              handleBlur={handleUpdateValue}
+            />
+          )}
         </p>
       </div>
-      <div className="opacity-0 group-hover:opacity-100 flex space-x-1 m-2">
-        {item.id && (
+      <div className="opacity-0 group-hover:opacity-100 absolute bottom-3 right-3 flex space-x-1">
+        {!isReadOnly && (
           <button
             className="text-red-600 bg-white w-8 h-8 rounded-full"
             onClick={() => handleDelete(item)}
@@ -99,7 +115,10 @@ function Input({ initialValue, type, handleBlur }) {
       value={value}
       onChange={(e) => setValue(e.target.value)}
       onBlur={() => onSubmit()}
-      className="bg-transparent font-semibold text-xl opacity-70 bg-blend-overlay text-black focus:text-gray-800 min-w-12 w-24 rounded outline-none px-1 text-right"
+      className="bg-transparent font-semibold text-xl bg-blend-overlay text-black focus:opacity-70 min-w-12 w-full rounded outline-none px-1 text-right appearance-none m-0"
+      style={{
+        marginRight: type === "number" && "-1rem",
+      }}
     />
   );
 }
